Moved Validations to the helper folder
